Moves CategorySelectionViewController dataSource func to CategorySelectionViewController+DataSource.swift file
