+++ quavi/quavi/Controllers/CategorySelectionViewController.swift
@@ -44,30 +44,3 @@ class CategorySelectionViewController: UIViewController {
         navigationController?.navigationBar.topItem?.title = "Tour Categories"
     }
 }
-
-extension CategorySelectionViewController: UICollectionViewDataSource{
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 10
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: Identifier.categoryCell.rawValue, for: indexPath) as? CatergoryCollectionViewCell else {return UICollectionViewCell()}
-        
-        cell.layer.borderWidth = 1
-        cell.layer.borderColor = #colorLiteral(red: 0.2843827307, green: 0.6391303539, blue: 0.8293711543, alpha: 1)
-        cell.layer.cornerRadius = 10
-        cell.clipsToBounds = true
-        return cell
-    }
-}
-
-extension CategorySelectionViewController: UICollectionViewDelegateFlowLayout{
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        
-        layout.sectionInset = UIEdgeInsets(top: 5, left: 5, bottom: 0, right: 5)
-        layout.minimumInteritemSpacing = 5
-        let virticalCellCGSize = CGSize(width: (collectionView.frame.size.width - 20), height: collectionView.frame.size.height / 6)
-        return virticalCellCGSize
-    }
-}
-
